# shared/api-contracts/ticketing-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Veraguas United Ticketing API
  version: 1.0.0
  description: >
    Contrato publico de Ticketing (Fase 7): eventos vendibles, ordenes,
    boletos y validacion en puerta. Los endpoints internos de operadores y
    escaner requieren autenticacion Sanctum (Bearer token).
  license:
    name: Propietaria
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1/ticketing
security: []
paths:
...
  /events/{publicId}/orders:
    post:
      summary: Crear orden (draft -> hold_active), reclama capacidad atomicamente
      operationId: createOrder
      parameters:
        - name: publicId
          in: path
          required: true
          schema: { type: string }
      responses:
        "201": { description: Orden creada, cupo reservado }
        "404": { description: Evento no encontrado }
        "409": { description: Cupo insuficiente o error de negocio }
...
components:
  securitySchemes:
    internalSecret:
      type: http
      scheme: bearer
    operatorAuth:
      type: http
      scheme: bearer
      description: Token Sanctum de operador obtenido via /auth/login